Added ActivePanels.au3
This tool helps align Panels in the Alignment Tool, and also allows to lock/unlock panel dragging, and serves as a proxy to activate dark mode or light mode from a custom powershell script. I also added a mini missing challenge into it, which will let Omnimo know if the file has been deleted by an antivirus as a false positive.